requirements.txt
streamlit
python-dotenv
numpy
transformers
langchain
langchain-community
langchain_google_genai
faiss-cpu